Name the nine Weapon nugget INI field parsers
The Weapon block's FieldParse table at 0x0109FE18 (read out of the exe,
docs/ini_schema.md) pairs each <Something>Nugget token with the address of
its INIFieldParseProc. Nine of those addresses were carried as Rva<hex>
symbols in Q4MakeNotifyLink.cpp, so a modder grepping AttributeModifierNugget
found nothing. Rename the file to WeaponNuggetParse.cpp and each entry point
to the key it parses, and type the two parameters INI* and WeaponTemplate*
now that the block parser (WeaponStore::parseWeaponTemplateDefinition) fixes
what the instance is.

The nugget classes, their constructors and their per-nugget field parsers are
pinned by address in reverse/symbols.csv and keep their address-derived names.
